requirements.txt
# ---------------------------------------------------------------------------
# Core transcription
# ---------------------------------------------------------------------------

# OpenAI Whisper (local model — no API key required)
# Installs torch as a dependency automatically.
openai-whisper>=20231117

# ---------------------------------------------------------------------------
# Audio processing
# ---------------------------------------------------------------------------

# pydub wraps ffmpeg for MP3→WAV conversion and audio splitting.
# ffmpeg must be installed separately (see README).
pydub>=0.25.1

# ---------------------------------------------------------------------------
# Web interface (optional — only needed for web_app.py)
# ---------------------------------------------------------------------------

flask>=3.0.0

# ---------------------------------------------------------------------------
# Indirect / pinned helpers
# ---------------------------------------------------------------------------

# numpy is required by Whisper for audio array processing
numpy>=1.24.0

# tiktoken is the tokeniser used by Whisper internally
tiktoken>=0.5.0

# more-itertools is a Whisper dependency
more-itertools>=9.0.0
